2.3 - Case Study Iteration 1 - Identifiable Object/IdenitfiableObject/IdentifiableObject/IdentifiableObject/IdentifiableObject.cs
using System;
using System.Collections;
using System.Collections.Generic;
using System.Linq;

namespace IdentifiableObject
{
    public class IdentifiableObject
    {
        private List<string> _identifiers = new List<string>();

        public IdentifiableObject(string[] idents)
        {
            foreach(string s in idents)
            {
                _identifiers.Add(s.ToLower());
            }
        }

        public bool AreYou(string id)
        {
            return _identifiers.Contains(id.ToLower());
        }

        public string FirstID
        {
            get
            {
                return _identifiers.First();
            }
        }

        public void AddIdentifier(string id)
        {
            _identifiers.Add(id.ToLower());
        }
    }
}

